pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cursorwise"
version = "0.1.0"
description = "MCP server for Cursor IDE <-> Flowise integration. 50 tools covering the full Flowise REST API."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
cursorwise = "cursorwise.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages]
find = {include = ["cursorwise", "cursorwise.*"]}
